--- menu_after_form/quiz/style_quiz.css
@@ -1,17 +1,0 @@
-@import "./media_queries_quizHTML/phone_screen.css" (min-width: 320px) and (max-width: 540px);
-@import "./media_queries_quizHTML/tablet_screen.css" (min-width: 540px) and (max-width: 768px);
-@import "./media_queries_quizHTML/pc_screen.css" (min-width: 768px);
-
-* {
-	box-sizing: border-box;
-	margin: 0;
-	padding: 0;
-}
-html {
-	height: 100%;
-}
-body {
-	background: linear-gradient(rgba(0, 0, 0, 0.704), rgb(255, 192, 203));
-	height: 100%;
-}
-
