+++ vite-project/package.json
@@ -1,0 +1,15 @@
+{
+	"name": "vite-project",
+	"private": true,
+	"version": "0.0.0",
+	"type": "module",
+	"scripts": {
+		"dev": "vite",
+		"build": "vite build",
+		"preview": "vite build && vite preview"
+	},
+	"devDependencies": {
+		"@types/node": "^18.11.18",
+		"vite": "^4.0.0"
+	}
+}
