+++ vite-project/src/Form/JavaScript/fetch_API_Weather.js
