+++ vite-project/src/Form/form.html
@@ -6,7 +6,6 @@
   <meta http-equiv="X-UA-Compatible" content="IE=edge" />
   <meta name="viewport" content="width=device-width, initial-scale=1.0" />
   <title>Document</title>
-  <link rel="stylesheet" href="style_form.css" />
   <script src="https://kit.fontawesome.com/3b117eb5e6.js" crossorigin="anonymous"></script>
   <link rel="preconnect" href="https://fonts.googleapis.com" />
   <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin />
@@ -79,12 +78,12 @@
         <p></p>
       </div>
       <p>click button and go to quiz</p>
-      <a href="./menu_after_form/menu.html"><button type="button" class="closeBtn">Close</button></a>
+      <a href="../Menu/menu.html"><button type="button" class="closeBtn">Close</button></a>
     </div>
 
   </div>
   <script src="https://cdn.jsdelivr.net/npm/axios/dist/axios.min.js"></script>
-  <script type="module" src="form.js"></script>
+  <script type="module" src="./main_form.js"></script>
 </body>
 
 </html>
