+++ vite-project/src/Form/main_form.js
@@ -1,0 +1,3 @@
+import "./Style/style_form.css";
+
+import "./JavaScript/form";
