+++ vite-project/src/Menu/JavaScript/display_nick_name.js
@@ -1,7 +1,8 @@
 const nickName = document.querySelector(".container-welcome-text h1");
-
 const displayNickName = () => {
 	nickName.textContent = localStorage.getItem("nickName");
 };
 
 displayNickName();
+
+export * from './display_nick_name.js';
