+++ vite-project/src/Menu/Style/media_queries_menuHTML/tablet_screen.css
@@ -1,3 +1,15 @@
+* {
+	box-sizing: border-box;
+}
+body {
+	height: 100%;
+	background: linear-gradient(rgb(0, 0, 3), rgb(255, 192, 203));
+	background-size: cover;
+	width: 100%;
+	margin: 0;
+	padding: 0;
+}
+
 .container-body {
 	display: flex;
 	justify-content: center;
