+++ vite-project/src/Menu/Style/menu_style.css
@@ -1,14 +1,4 @@
 @import "./media_queries_menuHTML/phone_screen.css" (min-width: 320px) and (max-width: 540px);
 @import "./media_queries_menuHTML/tablet_screen.css" (min-width: 540px) and (max-width: 768px);
 @import "./media_queries_menuHTML/pc_screen.css" (min-width: 768px);
-* {
-	box-sizing: border-box;
-}
-body {
-	height: 100%;
-	background: linear-gradient(rgb(0, 0, 3), rgb(255, 192, 203));
-	background-size: cover;
-	width: 100%;
-	margin: 0;
-	padding: 0;
-}
+
