+++ vite-project/src/Menu/main_menu.js
@@ -1,0 +1,6 @@
+import "./Style/menu_style.css";
+
+import "./JavaScript/weather_api";
+import "./JavaScript/select_quiz";
+import "./JavaScript/display_nick_name";
+import "./JavaScript/data_hour";
