+++ vite-project/src/Menu/menu.html
@@ -5,8 +5,7 @@
     <meta charset="UTF-8">
     <meta http-equiv="X-UA-Compatible" content="IE=edge">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Document</title>
-    <link rel="stylesheet" href="./menu_style.css">
+    <title>Menu</title>
     <script src="https://kit.fontawesome.com/3b117eb5e6.js" crossorigin="anonymous"></script>
 </head>
 
@@ -124,11 +123,8 @@
     </div>
 
     <script src="https://cdn.jsdelivr.net/npm/axios/dist/axios.min.js"></script>
-    <script type="module" src="./display_nick_name.js"></script>
-    <script type="module" src="./weather_api.js"></script>
-    <script type="module" src="./data_hour.js"></script>
-    <script type="module" src="./select_quiz.js"></script>
-
+    <script type="module" src="./main_menu.js"></script>
+    
 </body>
 
 </html>
