+++ vite-project/src/Quiz/Style/media_queries_quizHTML/phone_screen.css
@@ -1,0 +1,144 @@
+* {
+	box-sizing: border-box;
+	margin: 0;
+	padding: 0;
+}
+html {
+	height: 100%;
+}
+body {
+	background: linear-gradient(rgba(0, 0, 0, 0.704), rgb(255, 192, 203));
+	height: 100%;
+}
+
+.panel-before-quiz {
+	display: flex;
+	justify-content: center;
+	align-items: center;
+	text-align: center;
+	height: 70vh;
+}
+
+.start {
+	font-size: 40px;
+	font-weight: bold;
+	padding: 50px;
+	border-radius: 10px;
+	background-color: rgba(232, 162, 226, 0.323);
+	cursor: pointer;
+	outline: none;
+	border: 3px dashed black;
+	transition: background-color 0.5s;
+	margin-top: 10px;
+}
+
+.start:hover,
+.start:focus {
+	background-color: rgba(98, 57, 96, 0.646);
+}
+.instruction {
+	font-size: 40px;
+}
+
+.hidden-before-click-start {
+	display: none;
+	justify-content: center;
+	align-items: center;
+	height: 100%;
+}
+
+.body-forms {
+	border: 5px solid rgb(0, 0, 0);
+	width: 90%;
+	background-color: rgba(255, 255, 255, 0.564);
+}
+.header {
+	display: flex;
+	justify-content: space-around;
+	align-items: center;
+	margin-top: 5px;
+	font-size: 40px;
+}
+.question {
+	text-align: center;
+	font-size: 40px;
+	margin-top: 10px;
+}
+.all-answers {
+	margin: 20px;
+	display: flex;
+	flex-direction: column;
+}
+.all-answers button {
+	flex-wrap: wrap;
+	margin: 10px;
+	font-size: 15px;
+	padding: 10px;
+}
+.all-answers,
+.q {
+	transition: background-color 1s;
+}
+
+.q:hover,
+.q:focus {
+	background-color: rgb(139, 134, 134);
+}
+
+.body-button-next {
+	display: flex;
+	justify-content: flex-end;
+}
+.next-btn {
+	margin: 10px;
+	padding: 10px;
+	border-radius: 20px;
+	border: none;
+	background-color: rgba(0, 0, 0, 0.236);
+	transition: background 1s;
+	cursor: pointer;
+	outline: none;
+}
+.next-btn p {
+	font-size: 20px;
+}
+.next-btn:hover,
+.next-btn:focus {
+	background-color: rgb(146, 134, 134);
+}
+.final-score-body {
+	display: none;
+	justify-content: center;
+	align-items: center;
+	flex-direction: column;
+	height: 80%;
+}
+
+.final-score {
+	font-size: 40px;
+	color: rgb(251, 251, 250);
+	margin-bottom: 20px;
+}
+
+.end-buttons {
+	display: flex;
+	justify-content: space-around;
+	width: 90%;
+}
+
+.last-optionBtn {
+	background-color: rgba(0, 0, 0, 0.236);
+	padding: 20px 60px;
+	cursor: pointer;
+	border: none;
+	outline: none;
+	transition: background-color 1s;
+	border-radius: 5px;
+	font-size: 20px;
+	margin: 5px;
+}
+
+.last-optionBtn:hover,
+.last-optionBtn:focus {
+	background-color: rgba(216, 154, 194, 0.775);
+}
