+++ vite-project/src/Quiz/Style/style_quiz.css
@@ -1,0 +1,3 @@
+@import "./media_queries_quizHTML/phone_screen.css" (min-width: 320px) and (max-width: 540px);
+@import "./media_queries_quizHTML/tablet_screen.css" (min-width: 540px) and (max-width: 768px);
+@import "./media_queries_quizHTML/pc_screen.css" (min-width: 768px);
