+++ vite-project/src/Quiz/main_quiz.js
@@ -1,0 +1,3 @@
+import './Style/style_quiz.css'
+
+import './JavaScript/script_quiz'
