+++ vite-project/src/Quiz/quiz.html
@@ -5,8 +5,7 @@
   <meta charset="UTF-8" />
   <meta http-equiv="X-UA-Compatible" content="IE=edge" />
   <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-  <title>quiz</title>
-  <link rel="stylesheet" href="./style_quiz.css" />
+  <title>Quiz</title>
 </head>
 
 <body>
@@ -49,7 +48,7 @@
   <div class="final-score-body">
     <div class="final-score"></div>
     <div class="end-buttons">
-      <a href="../menu.html">
+      <a href="../Menu/menu.html">
         <button class="return last-optionBtn">Return</button>
       </a>
       <a href="./quiz.html">
@@ -58,9 +57,7 @@
     </div>
   </div>
 
-
-  <script type="module" src="../select_quiz.js"></script>
-  <script type="module" src="script_quiz.js" defer></script>
+  <script type="module" src="./main_quiz.js" defer></script>
 </body>
 
 </html>
